requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
google-generativeai==0.8.3
httpx==0.27.2
duckduckgo-search>=8.0.0
cachetools==5.4.0
pytest>=7.0.0